typescript mode updated

+++ api/src/controllers/vocabularyController.ts
@@ -1,32 +1,24 @@
-// vocabularyController.ts
 import { Request, Response } from "express";
-import Sentence from "../models/SuggestedSentence";
 import { AuthRequest } from "../middleware/authenticateToken";
-import { ObjectId } from "mongodb";
-import { Types, isValidObjectId } from "mongoose";
+import { Types } from "mongoose";
 import logger from "../utils/logger";
-import isValidObjectIdString from "../utils/isValidObjectIdString";
-import User from "../models/User";
-import updateRating from "../utils/updateRating";
-import Vocabulary from "../models/Vocabulary";
 import WordModel from "../models/Vocabulary/WordModel";
 import SuggestedWordModel from "../models/Vocabulary/SuggestedWordModel";
-import { validateSuggestedWordRequest } from "../middleware/validateSuggestedWord";
+import { ValidateSuggestedWordRequest } from "../middleware/validateSuggestedWord";
 
 const vocabularyController = {
   getAllWords: async (req: Request, res: Response) => {
     try {
-      // const wordsCount = await Vocabulary.find().countDocuments();
       const words = await WordModel.find()
         .sort({ _id: 1 })
         .populate("author", "_id firstname username email")
         .populate("translations", "_id text");
 
-      logger.info(`Все слова получены!`);
-      res.status(200).json({ message: `Словарь найден`, words });
+      logger.info("Все слова получены!");
+      res.status(200).json({ message: "Словарь найден", words });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Error retrieving words" });
+      logger.error(`Ошибка при получении слов: ${error}`);
+      res.status(500).json({ message: "Ошибка при получении слов" });
     }
   },
 
@@ -38,7 +30,7 @@ const vocabularyController = {
       const limitNumber = Number(limit);
       const skipIndex = (pageNumber - 1) * limitNumber;
 
-      const count = await SuggestedWordModel.find().countDocuments();
+      const count = await SuggestedWordModel.countDocuments();
       const words = await SuggestedWordModel.find()
         .sort({ _id: 1 })
         .skip(skipIndex)
@@ -46,100 +38,96 @@ const vocabularyController = {
         .populate("author", "_id firstname username email")
         .populate("pre_translations", "_id text");
 
-      logger.info(`Все предложенные слова получены!`);
+      logger.info("Все предложенные слова получены!");
       res
         .status(200)
-        .json({ message: `Словарь найден`, words, total_count: count });
+        .json({ message: "Словарь найден", words, total_count: count });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Error retrieving words" });
+      logger.error(`Ошибка при получении предложенных слов: ${error}`);
+      res
+        .status(500)
+        .json({ message: "Ошибка при получении предложенных слов" });
     }
   },
 
-  suggestWordTranslate: async (req: AuthRequest, res: Response) => {
+  suggestWordTranslate: async (
+    req: AuthRequest,
+    res: Response
+  ): Promise<Response> => {
     const { word_id, translate_language, translate, dialect } = req.body;
     try {
       if (!word_id || !translate_language || !translate) {
         logger.error(
-          `Ошибка при предложении перевода в словарь, отсутствует один из параметров`
+          "Ошибка при предложении перевода в словарь, отсутствует один из параметров"
         );
-        return res.status(400).json({
-          message: `Ошибка при предложении перевода в словарь, отсутствует один из параметров`,
-        });
+        return res
+          .status(400)
+          .json({ message: "Отсутствует один из параметров" });
       }
 
-      // Поиск исходного слова в базе данных
-      const is_exists = await WordModel.findById(word_id);
-      if (!is_exists) {
-        return res.status(400).json({
-          message: `Ошибка при предложении перевода в словарь, отсутствует исходное слово`,
-        });
+      const existingWord = await WordModel.findById(word_id);
+      if (!existingWord) {
+        return res.status(400).json({ message: "Исходное слово не найдено" });
       }
 
-      const is_exists_on_suggesteds = await SuggestedWordModel.findOne({
+      const existingSuggestedWord = await SuggestedWordModel.findOne({
         text: translate,
       });
-
-      if (is_exists_on_suggesteds) {
+      if (existingSuggestedWord) {
         await WordModel.findByIdAndUpdate(word_id, {
-          $addToSet: { translations: is_exists_on_suggesteds._id },
+          $addToSet: { translations: existingSuggestedWord._id },
         });
-        await SuggestedWordModel.findByIdAndUpdate(
-          is_exists_on_suggesteds._id,
-          {
-            $addToSet: { pre_translations: is_exists._id },
-          }
-        );
+        await SuggestedWordModel.findByIdAndUpdate(existingSuggestedWord._id, {
+          $addToSet: { pre_translations: existingWord._id },
+        });
       } else {
-        // Создание нового документа с переводом
-        const newSuggestedWord = await new SuggestedWordModel({
+        const newSuggestedWord = new SuggestedWordModel({
           text: translate,
           language: translate_language,
-          author: new Types.ObjectId(req.user.userId),
+          author: new Types.ObjectId(req.user?.userId),
           dialect,
-        }).save();
-        // Получение ID нового документа
-        const newSuggestedWordId = newSuggestedWord._id;
-        // Обновление поля translations в исходном слове
-        await WordModel.findByIdAndUpdate(word_id, {
-          $addToSet: { translations: newSuggestedWordId },
         });
-        // Обновление поля pre_translations в новом документе с переводом
-        await SuggestedWordModel.findByIdAndUpdate(newSuggestedWordId, {
-          $addToSet: { pre_translations: is_exists._id },
+        await newSuggestedWord.save();
+
+        await WordModel.findByIdAndUpdate(word_id, {
+          $addToSet: { translations: newSuggestedWord._id },
+        });
+        await SuggestedWordModel.findByIdAndUpdate(newSuggestedWord._id, {
+          $addToSet: { pre_translations: existingWord._id },
         });
       }
-      await is_exists.save();
 
+      await existingWord.save();
       return res
         .status(200)
         .json({ message: "Перевод успешно предложен и добавлен в словарь" });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ message: "Ошибка при предлоджении перевода" });
+      logger.error(`Ошибка при предложении перевода: ${error}`);
+      return res
+        .status(500)
+        .json({ message: "Ошибка при предложении перевода" });
     }
   },
 
   suggestWords: async (req: AuthRequest, res: Response) => {
     const { text, language } = req.body;
-
-    console.log(text)
-    console.log(language)
-
     try {
       if (!text || !language) {
         logger.error(
           "Ошибка при предложении слов, отсутствует один из параметров"
         );
-        return res.status(400).json({
-          message:
-            "Ошибка при предложении слов, отсутствует один из параметров",
-        });
+        return res
+          .status(400)
+          .json({ message: "Отсутствует один из параметров" });
       }
 
-      const wordsArray = text.split(",").map((word) => word.trim());
+      if (!req.user) {
+        return res.status(400).json({ message: "Отсутствует токен" });
+      }
+
       const userId = new Types.ObjectId(req.user.userId);
       const results = [];
+      const wordsArray = text.split(",").map((word: string) => word.trim());
 
       for (const word of wordsArray) {
         let existingWord = await SuggestedWordModel.findOne({ text: word });
@@ -155,13 +143,13 @@ const vocabularyController = {
             existingWord,
           });
         } else {
-          const newSuggestedWord = await new SuggestedWordModel({
+          const newSuggestedWord = new SuggestedWordModel({
             text: word,
             language,
             author: userId,
             contributors: [],
-          }).save();
-
+          });
+          await newSuggestedWord.save();
           results.push({
             message: "Слово успешно предложено",
             newSuggestedWord,
@@ -171,24 +159,28 @@ const vocabularyController = {
 
       return res.status(200).json(results);
     } catch (error) {
-      console.error(error);
+      logger.error(`Ошибка при предложении слов: ${error}`);
       res.status(500).json({ message: "Ошибка при предложении слов" });
     }
   },
 
   acceptSuggestedWord: async (
-    req: validateSuggestedWordRequest,
+    req: ValidateSuggestedWordRequest,
     res: Response
   ) => {
     const suggestedWord = req.suggestedWord;
 
     try {
+
+      if (!suggestedWord) {
+        return res.status(400).json({ message: 'suggestedWord не указан' })
+      }
+
       const wordToUpdate = await WordModel.findOne({
         text: suggestedWord.text,
       });
 
       if (wordToUpdate) {
-        // Обновление существующего документа
         wordToUpdate.language = suggestedWord.language;
         wordToUpdate.author = suggestedWord.author;
         wordToUpdate.contributors = Array.from(
@@ -202,7 +194,6 @@ const vocabularyController = {
         );
         await wordToUpdate.save();
       } else {
-        // Создание нового документа с сохранением старого ID
         await new WordModel({
           _id: suggestedWord._id,
           text: suggestedWord.text,
@@ -227,25 +218,25 @@ const vocabularyController = {
   },
 
   declineSuggestedWord: async (
-    req: validateSuggestedWordRequest,
+    req: ValidateSuggestedWordRequest,
     res: Response
   ) => {
     const suggestedWord = req.suggestedWord;
-
     try {
+      
+      if (!suggestedWord) {
+        return res.status(400).json({ message: 'SuggestedWord не указан' })
+      }
+      
       await SuggestedWordModel.findByIdAndDelete(suggestedWord._id);
-
       return res.status(200).json({ message: "Слово успешно отклонено" });
     } catch (error) {
-      console.error(error);
+      logger.error(`Ошибка при отклонении предложенного слова: ${error}`);
       res
         .status(500)
         .json({ message: "Ошибка при отклонении предложенного слова" });
     }
   },
-  // suggestWordTranslate: async (req: Request, res: Response) => {},
-  // suggestWordTranslate: async (req: Request, res: Response) => {},
-  // suggestWordTranslate: async (req: Request, res: Response) => {},
 };
 
 export default vocabularyController;
